requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
google-auth==2.24.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.96.0
sentence-transformers==2.2.2
chromadb==0.4.22
numpy==1.26.4
requests==2.31.0
jinja2==3.1.2
aiofiles==23.1.0
